backend/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
python-dotenv
pydantic
colorama
python-multipart
